move utils from bookmarkModule.js to bookmarkModule.utils.js
